backend/requirements.txt
llama-index-core
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-llms-nvidia 
llama-index-embeddings-nvidia
llama-index-utils-workflow
llama-index-postprocessor-rankgpt-rerank
llama-index-vector-stores-faiss
faiss-cpu
llama-index-readers-file
markdown_pdf
python-dotenv
pandas
fastapi
uvicorn
uvicorn[standard]
python-multipart
opentelemetry-exporter-otlp 
openinference-instrumentation-llama-index>=1.3.0
arize-phoenix